+++ perl/FSM/Support/NormalizedSemanticReport.pm
@@ -8,6 +8,7 @@ use JSON::PP ();
 
 use FSM::Support::CheckDiagnostics qw(build_check_failure_report build_check_success_report);
 use FSM::Support::CompositionReportContract qw(sanitize_composition_report);
+use FSM::Support::SerializableCompositionPlanSnapshot qw(build_serializable_composition_plan_snapshot);
 
 our @EXPORT_OK = qw(
     build_normalized_semantic_failure_report
@@ -131,6 +132,7 @@ sub _semantic_contract {
         target_language => $args{target_language},
     );
     my $composition = _composition_contract(
+        composition_plan => $result->{composition_plan},
         composition_report => $result->{composition_report} || $module_info->{composition_provenance},
         module_info => $module_info,
         intent_hir => $intent_hir,
@@ -238,6 +240,9 @@ sub _composition_contract {
     my $lowered_rtl_ir = $args{lowered_rtl_ir} || {};
     my $structural_rtl_ir = $args{structural_rtl_ir} || {};
     my $composition_report = sanitize_composition_report($args{composition_report});
+    my $composition_plan_snapshot = build_serializable_composition_plan_snapshot(
+        composition_plan => $args{composition_plan},
+    );
     my $is_composition = (
         exists $module_info->{composition_child_count}
         || exists $intent_hir->{composition_child_count}
@@ -294,6 +299,7 @@ sub _composition_contract {
                 || $lowered_rtl_ir->{composition_shared_datapath_candidates}
                 || []
         ),
+        plan_snapshot => $composition_plan_snapshot,
         (
             $composition_report
                 ? (provenance_report => $composition_report)